+++ base/exports.jl
@@ -997,6 +997,7 @@ export
     @polly,
 
     @assert,
+    @atomic,
     @__dot__,
     @enum,
     @label,