+++ NEWS.md
@@ -49,6 +49,10 @@ Command-line option changes
 Multi-threading changes
 -----------------------
 
+* Intrinsics for atomic pointer operations are now defined for certain byte sizes.
+* Support for declaring and using individual fields of a mutable struct as
+  atomic now available.
+
 
 Build system changes
 --------------------